#696 Charts: TreeViewer
refactoring
